requirements.txt
django==4.1.3
djangorestframework==3.14
mysql-connector-python==8.1.0
django-debug-toolbar==4.2.0
django-jazzmin==2.6.0
selenium==4.14.0
djangorestframework-api-key==3.0.0
django-environ==0.11.2

